consolidate rule filter functions
