+++ projects/rulesets-base/src/rule-runner/specification.ts
@@ -1,48 +1,12 @@
 import { OpenApiKind, OpenAPIV3, Result } from '@useoptic/openapi-utilities';
 
 import { createSpecification } from './data-constructors';
-import { RulesetData, NodeDetail } from './rule-runner-types';
-import {
-  createRulesetMatcher,
-  getRuleAliases,
-  createRuleContext,
-  isExempted,
-} from './utils';
+import { NodeDetail } from './rule-runner-types';
+import { createRuleContext, isExempted } from './utils';
 
 import { Rule, Ruleset, SpecificationRule } from '../rules';
 import { createSpecificationAssertions, AssertionResult } from './assertions';
-
-const getSpecificationRules = (
-  rules: (Ruleset | Rule)[]
-): (SpecificationRule & RulesetData)[] => {
-  const specificationRules: (SpecificationRule & RulesetData)[] = [];
-  for (const ruleOrRuleset of rules) {
-    if (SpecificationRule.isInstance(ruleOrRuleset)) {
-      specificationRules.push({
-        ...ruleOrRuleset,
-        aliases: [],
-      });
-    }
-
-    if (Ruleset.isInstance(ruleOrRuleset)) {
-      for (const rule of ruleOrRuleset.rules) {
-        if (SpecificationRule.isInstance(rule)) {
-          specificationRules.push({
-            ...rule,
-            matches: createRulesetMatcher({
-              ruleMatcher: rule.matches,
-              rulesetMatcher: ruleOrRuleset.matches,
-            }),
-            aliases: getRuleAliases(ruleOrRuleset.name, rule.name),
-            docsLink: rule.docsLink || ruleOrRuleset.docsLink,
-          });
-        }
-      }
-    }
-  }
-
-  return specificationRules;
-};
+import { getSpecificationRules } from './rule-filters';
 
 const createSpecificationResult = (
   assertionResult: AssertionResult,
